:recycle: refactor : style.ts 파일 분리
